move mock class into tests
