+++ src/simularium/index.tsx
@@ -1,6 +1,4 @@
-export type {
-    NetConnectionParams,
-} from "./RemoteSimulator";
+export type { NetConnectionParams } from "./RemoteSimulator";
 export type {
     VisDataMessage,
     VisDataFrame,
@@ -8,23 +6,10 @@ export type {
     EncodedTypeMapping,
     SimulariumFileFormat,
 } from "./types";
-export type {
-    SelectionStateInfo,
-    UIDisplayData
-} from "./SelectionInterface";
-export {
-    ErrorLevel, FrontEndError
-} from "./FrontEndError";
-export {
-    RemoteSimulator,
-    NetMessageEnum
-} from "./RemoteSimulator";
+export type { SelectionStateInfo, UIDisplayData } from "./SelectionInterface";
+export { ErrorLevel, FrontEndError } from "./FrontEndError";
+export { RemoteSimulator, NetMessageEnum } from "./RemoteSimulator";
 export { VisGeometry, RenderStyle, NO_AGENT } from "./VisGeometry";
 export { VisData } from "./VisData";
 export { ThreadUtil } from "./ThreadUtil";
-export {
-    SelectionInterface,
-} from "./SelectionInterface";
-
-export { DummyRemoteSimulator } from "./mock/DummyRemoteSimulator";
-
+export { SelectionInterface } from "./SelectionInterface";
